Correct export/import for interface
