+++ source/CellMeasurer/types.js
@@ -1,0 +1,15 @@
+// @flow
+
+export interface CellMeasureCache {
+  hasFixedWidth(): boolean;
+  hasFixedHeight(): boolean;
+  has(rowIndex: number, columnIndex: number): boolean;
+  set(
+    rowIndex: number,
+    columnIndex: number,
+    width: number,
+    height: number,
+  ): void;
+  getHeight(rowIndex: number, columnIndex?: number): number;
+  getWidth(rowIndex: number, columnIndex?: number): number;
+}
